+++ pages/dashboards/ocean-watch/index.jsx
@@ -1,9 +1,7 @@
 import PropTypes from 'prop-types';
 import dynamic from 'next/dynamic';
 import Link from 'next/link';
-import {
-  QueryClient,
-} from 'react-query';
+import { QueryClient } from 'react-query';
 import { dehydrate } from 'react-query/hydration';
 
 // components
@@ -15,17 +13,14 @@ import MapSelection from 'layout/layout/ocean-watch/map/component';
 import Banner from 'components/app/common/Banner';
 
 // services
-import {
-  fetchConfigFile,
-  fetchOceanWatchAreas,
-} from 'services/ocean-watch';
+import { fetchConfigFile, fetchOceanWatchAreas } from 'services/ocean-watch';
 
-const OceanWatchStoryTelling = dynamic(() => import('../../../layout/layout/ocean-watch/storytelling'), { ssr: false });
+const OceanWatchStoryTelling = dynamic(
+  () => import('../../../layout/layout/ocean-watch/storytelling'),
+  { ssr: false },
+);
 
-export default function OceanWatchIntroPage({
-  geostore,
-  oceanWatchConfig,
-}) {
+export default function OceanWatchIntroPage({ geostore, oceanWatchConfig }) {
   return (
     <LayoutOceanWatch
       title="Ocean Watch – Introduction"
@@ -46,38 +41,31 @@ export default function OceanWatchIntroPage({
           <section className="l-section -small">
             <div className="row">
               <div className="column small-12 medium-8">
-                <p style={{
-                  color: '#fff',
-                }}
-                >
-                  This is a Beta version of Ocean Watch -
-                  We&apos;re still tweaking things and making improvements.
-                  Please use the feedback button on the right to let us know
-                  if there is anything we can change to make your experience better!
+                <p className="text-white">
+                  This is a Beta version of Ocean Watch - We&apos;re still tweaking things and
+                  making improvements. Please use the feedback button on the right to let us know if
+                  there is anything we can change to make your experience better!
                 </p>
-                <p style={{
-                  color: '#fff',
-                }}
-                >
-                  The ocean and humanity are connected.
-                  To ensure the health and economic vitality of ocean ecosystems,
-                  ocean management needs an upgrade.
-                  Ocean Watch provides the data and information policymakers
-                  need to make better-informed decisions about sustainable ocean management.
+                <p className="text-white">
+                  The ocean and humanity are connected. To ensure the health and economic vitality
+                  of ocean ecosystems, ocean management needs an upgrade. Ocean Watch provides the
+                  data and information policymakers need to make better-informed decisions about
+                  sustainable ocean management.
                 </p>
               </div>
             </div>
           </section>
         </div>
-        <div style={{
-          position: 'absolute',
-          bottom: '0',
-          display: 'inline-block',
-          width: '100%',
-          height: 200,
-          background: 'linear-gradient(to bottom, transparent, #0F4573)',
-          pointerEvents: 'none',
-        }}
+        <div
+          style={{
+            position: 'absolute',
+            bottom: '0',
+            display: 'inline-block',
+            width: '100%',
+            height: 200,
+            background: 'linear-gradient(to bottom, transparent, #0F4573)',
+            pointerEvents: 'none',
+          }}
         />
       </div>
 
@@ -125,12 +113,10 @@ export default function OceanWatchIntroPage({
                   className="-text-center"
                   bgImage="/static/images/pages/app/banner-coral.jpg"
                 >
-                  <p className="-claim">
-                    Check the Coral Reefs dashboard
-                  </p>
+                  <p className="-claim">Check out the Coral Reefs dashboards</p>
                   <a
                     className="c-button -alt -primary"
-                    href="https://resourcewatch.org/dashboards/coral-reefs"
+                    href="https://resourcewatch.org/dashboards/coral-reef-dashboards"
                   >
                     Coral Reefs
                   </a>
@@ -156,18 +142,14 @@ export default function OceanWatchIntroPage({
                     }}
                   >
                     <div>
-                      <p className="-claim">
-                        Did you miss something?
-                      </p>
+                      <p className="-claim">Did you miss something?</p>
                       <p>
-                        Know of a data set that you&apos;d like to see on Resource Watch or have
-                        a specific area of interest you&apos;d like us to cover?
+                        Know of a data set that you&apos;d like to see on Resource Watch or have a
+                        specific area of interest you&apos;d like us to cover?
                       </p>
                     </div>
                     <Link href="/get-involved/contribute-data">
-                      <a className="c-button -alt -primary">
-                        Request data
-                      </a>
+                      <a className="c-button -alt -primary">Request data</a>
                     </Link>
                   </div>
                 </Banner>
@@ -195,12 +177,8 @@ export default function OceanWatchIntroPage({
                     }}
                   >
                     <div>
-                      <p className="-claim">
-                        What&apos;s Your Ocean Watch Story?
-                      </p>
-                      <p>
-                        How have you used Ocean Watch data to drive impact?
-                      </p>
+                      <p className="-claim">What&apos;s Your Ocean Watch Story?</p>
+                      <p>How have you used Ocean Watch data to drive impact?</p>
                     </div>
                     <a
                       className="c-button -alt -primary"
@@ -229,10 +207,8 @@ export default function OceanWatchIntroPage({
                     <br />
                     on the Explore page
                   </p>
-                  <Link href="/data/explore?section=All data&topics=[&quot;ocean&quot;]">
-                    <a className="c-button -alt -primary">
-                      Go to explore
-                    </a>
+                  <Link href='/data/explore?section=All data&topics=["ocean"]'>
+                    <a className="c-button -alt -primary">Go to explore</a>
                   </Link>
                 </Banner>
               </div>
@@ -263,14 +239,15 @@ export async function getServerSideProps() {
   await queryClient.prefetchQuery('ocean-watch-config-file', fetchConfigFile);
 
   // this page always uses a worldwide geostore
-  const { geostore } = queryClient.getQueryData('ocean-watch-areas').find(({ iso }) => iso === 'GLB') || {};
+  const { geostore } =
+    queryClient.getQueryData('ocean-watch-areas').find(({ iso }) => iso === 'GLB') || {};
 
   return {
-    props: ({
+    props: {
       dehydratedState: JSON.parse(JSON.stringify(dehydrate(queryClient))),
       geostore,
       oceanWatchConfig: queryClient.getQueryData('ocean-watch-config-file'),
-    }),
+    },
   };
 }